refactor to split out header for compile performance
