+++ pls.h
@@ -89,344 +89,59 @@ size_t find_dominant_ev(const EigenSolver<MATTYPE> es) {
 
 };
 
-float_type dominant_eigenvalue(const EigenSolver<Mat2Dc> es) {
-    const size_t idx = find_dominant_ev(es);
-    return abs(es.eigenvalues()[idx].real());
-};
-
-
-Colc dominant_eigenvector(const EigenSolver<Mat2D> es) {
-    const size_t idx = find_dominant_ev(es);
-    return es.eigenvectors().col(idx);
-}
-
-
-//
-// Numerical Approximation to Normal Cumulative Distribution Function
-//
-// DESCRIPTION:
-// REFERENCE: Handbook of Mathematical Functions with Formulas, Graphs, and Mathematical
-// Tables, U.S. Dept of Commerce - National Bureau of Standards, Editors: M. Abramowitz and I. A. Stegun
-// December 1972, p. 932
-// INPUT: z=computed Z-value
-// OUTPUT: probn=cumulative probability from -infinity to z
-//
-//
-float_type normalcdf(float_type z) {
-    const double c1 = 0.196854;
-    const double c2 = 0.115194;
-    const double c3 = 0.000344;
-    const double c4 = 0.019527;
-    float_type p;
-    if (z < 0) {
-        z = -z;
-        p = 1 - 0.5 / pow(1 + c1*z + c2*z*z + c3*z*z*z + c4*z*z*z*z, 4);
-    } else {
-        p = 0.5 / pow(1 + c1*z + c2*z*z + c3*z*z*z + c4*z*z*z*z, 4);
-    }
-    float_type probn = 1.0 - p;
-    return probn;
-}
-
-//
-// WILCOXON SIGNED RANK TEST FOR EVALUATING RELATIVE QUALITY OF TWO
-// COMPETING METHODS
-//
-// DESCRIPTION: Pairwise comparison between sets of model predictions
-// Competing models: model#1, model#2
-//
-// REFERENCE: Lehmann E. L. Nonparamtrics: Statistical Methods Based on Ranks.
-// Holden-Day: San Francisco, 1975, 120-132.
-//
-// Let: U=sum of postive ranks, V=sum of negative ranks
-// (V>U) is evidence that the model#1 is better)
-// Define: d=U-V and t=U+V=n(n+1)/2
-// Then V=(t-d)/2
-//
-// Asymptotic Theory: Suppose n is the number of samples.
-// Then, E(V)=n(n+1)/4 and Var(V)=n(n+1)(2n+1)/24.
-// It follows that (V-E(V))/Std(V) is approx. normally distributed.
-//
-// INPUT: err_1=prediction errors from model#1
-//        err_2=prediction errors from model#2
-//
-// OUTPUT: probw=Prob{V is larger than observed}
-// If probw is small enough, conclude that model#1 is better
-//
-// Based on Matlab code from
-// Thomas E. V. Non-parametric statistical methods for multivariate calibration
-// model selection and comparison. J. Chemometrics 2003; 17: 653–659
-//
-float_type wilcoxon(const Col err_1, const Col err_2) {
-    assert(err_1.rows() == err_2.rows());
-
-    const size_t n = err_1.rows();
-    Col del = err_1.cwiseAbs() - err_2.cwiseAbs();
-    assert(static_cast<size_t>(del.size()) == n);
-
-    // TODO: setZero necessary?
-    Rowi sdel;
-    sdel.setZero(n);
-    //Matrix<int, Dynamic, 1> sdel = del.unaryExpr(std::ptr_fun(_sgn)); // can't get this to work
-    for (size_t i = 0; i < n; i++) {
-        sdel(i) = (0 < del(i)) - (del(i) < 0); // get the sign of each element
-    }
-    Col adel = del.cwiseAbs();
-    // 's' gives the original positions (indices) of the sorted values
-    auto s = ordered(adel);
-    float_type d = 0;
-    for (size_t i = 0; i < n; i++) { d += (i + 1) * sdel(s[i]); }
-    float_type t  = n * (n + 1) / 2.0;
-    float_type v  = (t - d) / 2.0;
-    float_type ev = t/2.0;
-    float_type sv = sqrt(static_cast<float_type>(n * (n+1) * (2*n+1)) / 24.0);
-    float_type z = (v - ev) / sv;
-    float_type probw = 1.0 - normalcdf(z);
-
-    return probw;
-}
-
-// TODO: several of the loop constructs seem ripe for row/col-wise operations / broadcasting:
-// https://eigen.tuxfamily.org/dox/group__TutorialReductionsVisitorsBroadcasting.html
-
 struct PLS_Model {
 
-    Mat2Dc P, W, R, Q, T;
-    size_t A;
-    METHOD method;
-    void initialize(size_t num_predictors, size_t num_responses, size_t num_components) {
-        A = num_components;
-        P.setZero(num_predictors, num_components);
-        W.setZero(num_predictors, num_components);
-        R.setZero(num_predictors, num_components);
-        Q.setZero(num_responses, num_components);
+    PLS_Model(
+      const size_t num_predictors, const size_t num_responses, const size_t num_components
+    ) : A(num_components), P(num_predictors, num_components), W(num_predictors, num_components),
+        R(num_predictors, num_components), Q(num_responses, num_components) {
         // T will be initialized if needed
-        return;
     }
 
-    //"Modified kernel algorithms 1 and 2"
-    //from Dayal and MacGregor (1997) "Improved PLS Algorithms" J. of Chemometrics. 11,73-85.
-    void plsr(const Mat2D X, const Mat2D Y, METHOD algorithm) {
-        method = algorithm;
-        int M = Y.cols(); // Number of response variables == columns in Y
-
-        if (algorithm == KERNEL_TYPE1) T.setZero(X.rows(), A);
-
-        Mat2D XY = X.transpose() * Y;
-        Mat2D XX;
-        if (algorithm == KERNEL_TYPE2) XX = X.transpose() * X;
-
-        for (size_t i = 0; i < A; i++) {
-            Colc w, p, q, r, t;
-            complex<float_type> tt;
-            if (M == 1) {
-                w = XY.cast<complex<float_type> >();
-            } else {
-                EigenSolver<Mat2D> es( (XY.transpose() * XY) );
-                q = dominant_eigenvector(es);
-                w = (XY*q);
-            }
-
-            w /= sqrt((w.transpose()*w)(0,0)); // use normalize function from eigen?
-            r = w;
-            for (size_t j=0; j <= i-1; j++) {
-                r -= (P.col(j).transpose()*w)(0,0)*R.col(j);
-            }
-            if (algorithm == KERNEL_TYPE1) {
-                t = X*r;
-                tt = (t.transpose()*t)(0,0);
-                p.noalias() = (X.transpose()*t);
-            } else if (algorithm == KERNEL_TYPE2) {
-                tt = (r.transpose()*XX*r)(0,0);
-                p.noalias() = (r.transpose()*XX).transpose();
-            }
-            p /= tt;
-            q.noalias() = (r.transpose()*XY).transpose(); q /= tt;
-            XY -= ((p*q.transpose())*tt).real(); // is casting this to 'real' always safe?
-            W.col(i) = w;
-            P.col(i) = p;
-            Q.col(i) = q;
-            R.col(i) = r;
-            if (algorithm == KERNEL_TYPE1) T.col(i) = t;
-        }
-        return;
-    }
-
+    void plsr (const Mat2D& X, const Mat2D& Y, const METHOD algorithm);
 
     // latent X values, i.e. the orthogonal metrics you wish you could measure
+    const Mat2Dc scores(const Mat2D& X_new, const size_t comp) const;
     const Mat2Dc scores(const Mat2D& X_new) const { return scores(X_new, A); }
-    const Mat2Dc scores(const Mat2D& X_new, const size_t comp) const {
-        assert (A >= comp);
-        return X_new * R.leftCols(comp);
-    }
 
     // compute the regression coefficients (aka 'beta')
+    const Mat2Dc coefficients(const size_t comp) const;
     const Mat2Dc coefficients() const { return coefficients(A); }
-    const Mat2Dc coefficients(const size_t comp) const {
-        assert (A >= comp);
-        return R.leftCols(comp)*Q.leftCols(comp).transpose();
-    }
+
 
     // predicted Y values, given X values and pls model
+    const Mat2D fitted_values(const Mat2D& X, const size_t comp) const;
     const Mat2D fitted_values(const Mat2D& X) const { return fitted_values(X, A); }
-    const Mat2D fitted_values(const Mat2D& X, const size_t comp) const {
-        assert (A >= comp);
-        return X*coefficients(comp).real();
-    }
 
     // unexplained portion of Y values
+    const Mat2D residuals(const Mat2D& X, const Mat2D& Y, const size_t comp) const;
     const Mat2D residuals(const Mat2D& X, const Mat2D& Y) const { return residuals(X, Y, A); }
-    const Mat2D residuals(const Mat2D& X, const Mat2D& Y, const size_t comp) const {
-        assert (A >= comp);
-        return Y - fitted_values(X, comp);
-    }
 
     // Sum of squared errors
+    const Row SSE(const Mat2D& X, const Mat2D& Y, const size_t comp) const;
     const Row SSE(const Mat2D& X, const Mat2D& Y) const { return SSE(X, Y, A); }
-    const Row SSE(const Mat2D& X, const Mat2D& Y, const size_t comp) const {
-        return residuals(X, Y, comp).colwise().squaredNorm();
-    }
 
     // Total sum of squares
-    Row SST(const Mat2D& Y) const {
-        Row sst(Y.cols());
-        for (size_t c = 0; c < static_cast<size_t>(Y.cols()); c++) {
-            sst(c) = (Y.col(c).array() - (Y.col(c).sum()/Y.rows())).square().sum();
-        }
-        return sst;
-    }
+    Row SST(const Mat2D& Y) const;
 
     // fraction of explainable variance
+    Row explained_variance(const Mat2D& X, const Mat2D& Y, const size_t comp) const;
     Row explained_variance(const Mat2D& X, const Mat2D& Y) const { return explained_variance(X, Y, A); }
-    Row explained_variance(const Mat2D& X, const Mat2D& Y, const size_t comp) const {
-        assert (A >= comp);
-        return (1.0 - SSE(X, Y, comp).cwiseQuotient( SST(Y) ).array()).matrix();
-    }
 
     // leave-one-out validation of model (i.e., are we overfitting?)
-    Mat2D loo_validation(const Mat2D& X, const Mat2D& Y, const VALIDATION_OUTPUT out_type) const {
-        Mat2D Xv = X.bottomRows(X.rows()-1);
-        Mat2D Yv = Y.bottomRows(Y.rows()-1);
+    Mat2D loo_validation(const Mat2D& X, const Mat2D& Y, const VALIDATION_OUTPUT out_type) const;
 
-        Mat2D SSEv = Mat2D::Zero(Y.cols(), this->A);
-
-        PLS_Model plsm_v;
-        plsm_v.initialize(Xv.cols(), Yv.cols(), this->A);
-        for (size_t i = 0; i < static_cast<size_t>(X.rows())-1; i++) {
-            // run pls for the data, less one observation
-            plsm_v.plsr(Xv, Yv, this->method);
-            for (size_t j = 1; j <= this->A; j++) {
-                // now see how well the data predict the missing observation
-                Row res = plsm_v.residuals(X.row(i), Y.row(i), j).row(0);
-                // tally the squared errors
-                SSEv.col(j - 1) += res.cwiseProduct(res).transpose();
-            }
-            Xv.row(i) = X.row(i);
-            Yv.row(i) = Y.row(i);
-        }
-        if ( out_type == PRESS ) {
-            return SSEv;
-        } else if ( out_type == RMSEP ) {
-            SSEv /= X.rows();
-            return SSEv.cwiseSqrt();
-        }
-    }
-
-
-    std::vector<Mat2D> _loo_cv_error_matrix(const Mat2D& X, const Mat2D& Y) const {
-        Mat2D Xv = X.bottomRows(X.rows()-1);
-        Mat2D Yv = Y.bottomRows(Y.rows()-1);
-
-        // vector of error matrices(rows=Y.rows(), cols=Y.cols())
-        // col = component #, row = obs #, tier = Y category
-        std::vector<Mat2D> Ev(Y.cols(), Mat2D::Zero(X.rows(), this->A));
-
-        PLS_Model plsm_v;
-        plsm_v.initialize(Xv.cols(), Yv.cols(), this->A);
-        for (size_t i = 0; i < static_cast<size_t>(X.rows()); i++) {
-            plsm_v.plsr(Xv, Yv, this->method);
-            for (size_t j = 1; j <= this->A; j++) {
-                Row res = plsm_v.residuals(X.row(i), Y.row(i), j).row(0);
-                for (int k = 0; k < res.size(); k++) Ev[k](i,j-1) = res(k);
-            }
-            if (i < static_cast<size_t>(Xv.rows())) {
-                Xv.row(i) = X.row(i);
-                Yv.row(i) = Y.row(i);
-            }
-        }
-        return Ev;
-    }
-
-
-    std::vector<Mat2D> _new_data_cv_error_matrix(const Mat2D& X_new, const Mat2D& Y_new) const {
-        // vector of error matrices(rows=Y.rows(), cols=Y.cols())
-        // col = component #, row = obs #, tier = Y category
-        std::vector<Mat2D> Ev(Y_new.cols(), Mat2D::Zero(X_new.rows(), this->A));
-
-        for (size_t j = 1; j <= this->A; j++) { // j is component #
-            Mat2D res = residuals(X_new, Y_new, j);
-            for (size_t k = 0; k < static_cast<size_t>(res.cols()); k++) { // k is Y category
-                Ev[k].col(j - 1) = res.col(k);
-            }
-        }
-        return Ev;
-    }
+    std::vector<Mat2D> _loo_cv_error_matrix(const Mat2D& X, const Mat2D& Y) const;
+    std::vector<Mat2D> _new_data_cv_error_matrix(const Mat2D& X_new, const Mat2D& Y_new) const;
 
     // if val_method is LOO, X and Y should be original data
     // if val_method is NEW_DATA, X and Y should be observations not included in the original model
-    Rowsz optimal_num_components(const Mat2D& X, const Mat2D& Y, const VALIDATION_METHOD val_method) const {
-        // col = component #, row = obs #, tier = Y category
+    const Rowsz optimal_num_components(const Mat2D& X, const Mat2D& Y, const VALIDATION_METHOD val_method) const;
 
-        std::vector<Mat2D> errors;
-        if (val_method == LOO) {
-            errors = _loo_cv_error_matrix(X, Y);
-        } else {
-            errors = _new_data_cv_error_matrix(X, Y);
-        }
-
-        Mat2D press = Mat2D::Zero(Y.cols(), A);
-        Rowsz min_press_idx = Rowsz::Zero(Y.cols());
-        Row  min_press_val(Y.cols());
-        Rowsz best_comp(Y.cols());
-
-        // Determine PRESS values
-        for (size_t i = 0; i < static_cast<size_t>(errors.size()); i++) {    // for each Y category
-            for (size_t j = 0; j < static_cast<size_t>(errors[i].rows()); j++) {      // for each observation
-                for (size_t k = 0; k < static_cast<size_t>(errors[i].cols()); k++) {  // for each component
-                    press(i,k) += pow(errors[i](j,k), 2);
-                }
-            }
-        }
-
-        min_press_val = press.col(0);
-        // Find the component number that minimizes PRESS for each Y category
-        for (size_t i = 0; i < static_cast<size_t>(press.rows()); i++) {              // for each Y category
-            for (size_t j = 0; j < static_cast<size_t>(press.cols()); j++) {          // for each component
-                if (press(i,j) < min_press_val(i)) {
-                    min_press_val(i) = press(i,j);
-                    min_press_idx(i) = j;
-                }
-            }
-        }
-
-        best_comp = min_press_idx.array() + 1; // +1 to convert from index to component number
-        // Find the min number of components that is not significantly
-        // different from the min PRESS at alpha = 0.1 for each Y category
-        const float_type ALPHA = 0.1;
-        for (size_t i = 0; i < static_cast<size_t>(press.rows()); i++) {              // for each Y category
-            for (size_t j = 0; j < min_press_idx(i); j++) {      // for each smaller number of components
-                Col err1 = errors[i].col(min_press_idx(i));
-                Col err2 = errors[i].col(j);
-                auto p = wilcoxon(err1, err2);
-                if (p > ALPHA) {
-                    best_comp(i) = j + 1; // +1 to convert from index to component number
-                    break;
-                }
-            }
-        }
-
-        return best_comp;
-    }
+    private:
+        size_t A; // number of components
+        Mat2Dc P, W, R, Q, T;
+        METHOD method;
 
 };
